+++ requirements/frontend/src/Routes.tsx
@@ -13,6 +13,7 @@ import { Lobby } from './pages/lobby/Lobby';
 import { Login } from './pages/login/Login';
 import { RoomChecker } from './pages/channel/components/RoomChecker';
 import { ErrorHandler } from './ErrorHandler';
+import { Modal } from './modal/Modal';
 
 export const Routes = createBrowserRouter(
   createRoutesFromElements(
@@ -28,10 +29,12 @@ export const Routes = createBrowserRouter(
       <Route index element={<Navigate to='/login' replace={true} />} />
       <Route path='login' element={<Login />} />
       <Route element={<LoginChecker />}>
-        <Route path='lobby' element={<Lobby />} />
-        <Route path='channel' element={<Channel />} />
-        <Route path='channel/room' element={<Room />} />
-        <Route path='channel/roomChecker' element={<RoomChecker />} />
+        <Route element={<Modal />}>
+          <Route path='lobby' element={<Lobby />} />
+          <Route path='channel' element={<Channel />} />
+          <Route path='channel/room' element={<Room />} />
+          <Route path='channel/roomChecker' element={<RoomChecker />} />
+        </Route>
       </Route>
     </Route>,
   ),
