+++ requirements/frontend/src/atoms/modals/newRoomModalState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const newRoomModalState = atom<boolean>({
+  key: 'newRoomModalState',
+  default: false,
+});
