+++ requirements/frontend/src/components/SideBar/RankList.tsx
@@ -6,7 +6,7 @@ import { useInterval } from '../../hooks/useInterval';
 import { UserDataType } from '../../atoms/userDataType';
 import styles from './Li/UserLi.module.css';
 import { useSetRecoilState } from 'recoil';
-import { userProfileModalState } from '../../atoms/userProfileModalState';
+import { userProfileModalState } from '../../atoms/modals/userProfileModalState';
 
 export function RankList() {
   const { token } = useRecoilValue(authState);
