+++ requirements/frontend/src/modal/Modal.tsx
@@ -1,17 +1,17 @@
+import { Outlet } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
-import { userProfileModalState } from '../atoms/userProfileModalState';
-import { UserProfile } from './components/UserProfileModal';
+import { userProfileModalState } from '../atoms/modals/userProfileModalState';
+import { NewRoomModal } from './components/NewRoomModal';
+import { UserProfileModal } from './components/UserProfileModal';
 
 export function Modal() {
   const userProfileModal = useRecoilValue(userProfileModalState);
 
   return (
     <>
-      {userProfileModal === undefined ? (
-        <></>
-      ) : (
-        <UserProfile id={userProfileModal} />
-      )}
+      <UserProfileModal />
+      <NewRoomModal />
+      <Outlet />
     </>
   );
 }
