+++ requirements/frontend/src/modal/components/Modal.module.css
@@ -1,0 +1,11 @@
+.modal {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  top: 0px;
+  left: 0px;
+  width: 1280px;
+  height: 720px;
+  /* background-color: rgba(0, 0, 0, 0.3); */
+}
