+++ requirements/frontend/src/modal/components/UserProfileModal.module.css
@@ -1,15 +1,3 @@
-.modal {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  width: 1280px;
-  height: 720px;
-  /* background-color: rgba(0, 0, 0, 0.3); */
-}
-
 .profile__blank {
   z-index: 49;
   position: absolute;
